+++ rust/docs/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to the **rust** extension will be documented in this file.
 
+## [1.20.5] - 2026-05-30
+
+### Fixed
+- keep runner prelude self-contained
+
+## [1.20.4] - 2026-05-30
+
+### Changed
+- use typed sidecar writer APIs
+
+## [1.20.3] - 2026-05-30
+
+### Changed
+- Use shared runner prelude in extension runners
+
 ## [1.20.2] - 2026-05-30
 
 ### Changed